Improved code readability, included default schema.yml filename, modulate to move to async.
